requirements.txt
torch>=1.7.0
torchvision>=0.8.0
torchaudio>=0.7.0
timm>=0.4.12
numpy>=1.19.0
matplotlib>=3.3.0
pillow>=8.0.0
librosa>=0.8.0
sounddevice>=0.4.1
soundfile>=0.10.3
scipy>=1.5.0
pyinstaller>=4.5.0